+++ R/packrat.R
@@ -254,7 +254,7 @@ initImpl <- function(project = getwd(),
 
     # Restart R if the environment is capable of it (otherwise enter packrat mode)
     if (!restart || !attemptRestart())
-      on(project = project, clean.search.path = TRUE)
+      on(project = project, clean.search.path = FALSE)
   }
 
   invisible()