+++ NEWS.md
@@ -39,6 +39,9 @@
 - Infer package dependencies from `requireNamespace()` and `loadNamespace()`
   only when the package name is character input. (#602)
 
+- Allow repository names with commas. Repository URLs containing commas are
+  still not supported. (#587)
+
 # Packrat 0.5.0
 
 - Packrat now supports both of BiocManager and BiocInstaller (as used for